Split load balancing module into three submodules
Split load balancing module into 3 submodules, each corresponding to
one policy. Refactored tests, removed code duplication by creating
constructor to ClusterData struct (needed for test mocks).
